+++ api/internal/service/measurement.go
@@ -38,7 +38,7 @@ type noteCbk func(context.Context, uuid.UUID, time.Time, model.TimeseriesNote) e
 func createMeasurements(ctx context.Context, mc []model.MeasurementCollection, mmtFn mmtCbk, noteFn noteCbk) error {
 	for _, c := range mc {
 		for _, m := range c.Items {
-			if err := mmtFn(ctx, c.TimeseriesID, m.Time, m.Value); err != nil {
+			if err := mmtFn(ctx, c.TimeseriesID, m.Time, float64(m.Value)); err != nil {
 				return err
 			}
 			if m.Masked != nil || m.Validated != nil || m.Annotation != nil {
